Move signup code to new location
